Allow 'salt' command metadata to be a dict
When invoking 'salt' on the command line, using the '--metadata'
option will not allow you to pass in data that will be represented
in Python as a dict. This data will be represented as a string.
There are situations where using a dict for metadata is desirable.
For instance, if one would want to pass a number of named options as
metadata.

To accomplish this, the metadata is simply fed through the function:
'salt.utils.args.yamlify_arg'. This function is versatile: if the
original input is not a string, it will return the original input. If
it fails to parse the input into the dict or other such object, it
will return the original input.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>

Conflicts:
	salt/cli/salt.py

Conflicts:
	salt/cli/salt.py