Minimize DNS resolution queries on grains load
When grains are loaded, hostname() is called multiple times. The
hostname() function ends up calling socket.getfqdn, which is an
expensive operation if the host is not on the domain (the function times
out). Optimize grains operation to only query getfqdn once. Also , optimize
fqdn_ip4 and fqdn_ipv6 grains to only call socket.getaddrinfo (expensive
operation) if there is a domain.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>