Fix logging based memory leaks on Windows for 2015.8
Did not pull in these commits:

Fixed Windows salt-master memory leaks in logger
793c0bef75bf6c2700b8f041d57fdad5d9b5e1bd

Fix more logger related memory leaks on Windows
94bb66f2f7b7f07edf18d889e528fd93c2ee5eb5

Due to having a merge nightmare even after trying to cherry-pick
about 30 commits. Instead, solved the problem a different way.

The original solution in salt/develop is more robust, but this
solution should suit our needs for 2015.8.

This solution does the following:
On Windows, if it is not the main process, ensure that the temp null and
temp queue logging handlers are not installed when salt/log/setup.py is
loaded.

This solution will NOT be upstreamed and will be obsoleted when we move
to the next version of Salt.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>