Fix TCP transport presence events to work with a NAT
Currently, the presence events do not work when a salt-master is
communicating with salt-minions that are behind a NAT. This is a common
cloud administration use case. It doesn't work for the following reasons:

- The IP address that the salt-master sees is the NAT IP address and it
is different than the IP address that the salt-minion thinks it has.
Hence matching by IP address fails.
- The salt-master may see multiple minions through a NAT with the same
IP address, but different port numbers.

The following challenges have been observed:
- The PUB server connections are the ones that are persistent.
Unfortunately, the server has no idea which PUB connection corresponds
to which salt-minion.
- The salt-master can figure out which salt-minion it is communicating
with it over the REQ server connections. However, these are not
persistent and there is no way to correlate the REQ Server connections
with the PUB server connections. The reason is that in the NAT case,
multiple salt-minions can have the same IP address. And the REQ and PUB
port numbers will obviously be different.

Due to these challenges, the following solution has been implemented:
- When the minion connects or reconnects the PUB connection to the
master, it will send data that identifies it to the master over the PUB
connection. This data is encrypted with AES and includes an identifying
token, similar to how REQ connections are protected.
- Once the master receives this data and verifies its authenticity, it
will proceed to fire the correct presence events. Also, when the
connection is disconnected, it will also proceed to fire the correct
presence events.
- These presence events will fire from the PUB Server only when the TCP
transport is the only transport used, and when `presence_events` is
True in the opts. In this case, it will disable the `presence_events`
functionality in the `Maintenance` process.

Additional comments:
- Now we also selectively publish to minions (when using tgt_type
'list') instead of always publishing to all connected minions, since
we now associate a minion id for each PUB connection. For large numbers
of minions, publishing to all minions when only a small subset are
needed greatly increase network traffic and degrade performance.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>

Conflicts:
	salt/master.py