Fix TCP transport PUB auth on failover with master restart
If the salt-master restarts, and the salt-minion is using the TCP
transport on failover mode, the salt-master will reject the
salt-minion's pre-existing auth token.

This fix will reset the auth when handling the `__master_disconnected`
event in minion.py.

Note: A possible alternative to this fix is to have the master respond
to the minion during the PUB channel identification process. If the
response is that the authentication failed, the minion would then try
to re-authenticate at that time. Currently, the master does not
respond to the minion's identification token.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>