Fixed error when first job is multi-func
*** Not for upstreaming ***

When first job is multi-func, an error such as this appears:

[WARNING ] The minion function caused an exception: 'Minion' object
has no attribute 'functions'

And jobs are not properly completed. The reason is that
`_thread_multi_return` doesn't fully initialize `minion_instance`
like `_thread_return` does. This just copies the exact same code
from `_thread_return` into `_thread_multi_return`.

This is not an issue in 2016.3 because they re-architected this code
in a way where this initialization is done before `_thread_return`
and `_thread_multi_return`. Due to this, this change should not be
upstreamed.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>