Do not exit process on invalid master key
Currently, the behavior is to exit the salt-minion process on
an invalid master key. This change will pass an error back via
'SaltClientError' instead. Some benefits of this:

- If 'failover' mode is in use, it will correctly move to the
next master in the list.
- Relying on the salt-minion keep-alive mechanism to retry
may not be desirable (extra process) or available (not supported
on Windows minions).
- This will defer to the pre-existing mechanisms of `auth_tries`
and `master_tries` which seem better tailored for this type of
issue.
- There has been observed a race condition (at least in the TCP
transport) where if the salt-master goes up at a certain time when
the salt-minion is already trying to connect to the master, the
'invalid master key' condition will be triggered. This condition will
automatically clear itself on the next retry / failover.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>