Add 'resolve_dns_fallback' option to config
By default, the salt-minion will fall back to localhost (127.0.0.1) if
the DNS resolve fails on the salt-master hostname. There are some cases
where that is not desirable, such as when using a master failover list
and there is a master running locally that we don't want to connect to.

Add a boolean config option called `resolve_dns_fallback` which
defaults to True but will disable the fall back behavior if set to
False.

Now that `salt.minion.resolve_dns()` may be invoked with `fallback=False`,
that function is more likely to throw a `SaltClientError` exception
instead of reacting to the error condition internally. Due to this, made
sure the callers will be able to catch that exception and react
appropriately.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>