opkg module: Add 'version', 'reinstall', and 'only_upgrade' to install (#37008)
Add support for the options `version`, `reinstall`, and `only_upgrade`
to the `install()` function of the `opkg` module. These options reflect
the behavior of the same options in the `aptpkg` module.

This also adds support to using a format for the `pkgs` option that looks
like:

```
pkgs='["foo", {"bar": "1.2.3-0ubuntu0"}]'
```

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>