Add config options for running beacons/scheduler before connect
Add the following configuration options to tell a minion to
set up its beacons and/or scheduler before a connection to the master
has been made: `beacons_before_connect`, `scheduler_before_connect`.

An example where this option would be useful is if we are using the
avahi_announce or bonjour_announce beacon and we would like mDNS
announcements to occur before connect, so that there is always an
announcement even if the master is down or an incorrect master hostname
is configured.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>