Fix Current-Gen detection when run in daemon
Current-Gen is not properly detected with
`salt.utils.which('nisafemodeversion')`. It will work when not running
in a daemon, for example when using `salt-call`. But when the salt-master
calls into the salt-minion, the daemon environment doesn't have
`/usr/local/natinst/bin` in the path, so the call will fail.
Use the complete file path to check if it exists.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>

Conflicts:
	salt/modules/timezone.py