Windows: Fix usage of pkgrepo state
`pkgrepo.py` contains this line:

```
from salt.modules.aptpkg import _strip_uri
```

However, `salt.modules.aptpkg` is just not importable on Windows.
Instead of trying to make this module importable on Windows (which
clearly isn't the intent of the module), placed a copy of `_strip_uri`
into `pkgrepo.py` as this is a small function.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>