Remove a specific `log.debug` that causes an MWorker hang
There is a specific `log.debug` statement that causes a hang in an
MWorker after a `Salt request timed out. The master is not responding.`
exception. It doesn't occur every time, but once it does, it makes
the MWorker unusable until the salt-master is restarted. The cause was
traced to `salt/log/handlers/__init__.py`, the `QueueHandler` object
would block in this call:

```
self.queue.put_nowait(record)
```

Whether we use the old Python 2 `QueueHandler` or the Python 3
`logging.handlers.QueueHandler`, the result is the same block. We
currently have no idea why this blocks, as it behaves contrary to
the Python documentation for a multiprocessing queue:
https://docs.python.org/3.6/library/multiprocessing.html

It may be a race condition in the Python implementation that we
currently use, but we are not sure. It likely only occurs on Windows
salt-master, so the Salt community hasn't seen it.

We don't understand why only this particular `log.debug` causes the
issue. This commit is to make the issue go away until we can address
it in a more robust manner.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>