opkg: add minion reboot support
Some packages on NILRT systems, especially kernel modules which are a
poignant exception in the Linux world for this use case, require
rebooting the minion after an opkg install/remove/upgrade/etc.

This adds support for tracking kernel module states to know when a
system reboot is necessary (via restartcheck) and calls a 'witness'
function to set a temporary 'reboot required' state which is cleared
on boot.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>